moved forms into new 'mini_training.py' file
