+++ base/views/education_groups/update.py
@@ -29,8 +29,8 @@ from django.shortcuts import get_object_or_404, redirect
 from django.utils.translation import ugettext_lazy as _
 from waffle.decorators import waffle_flag
 
-from base.forms.education_group.create import CreateEducationGroupYearForm, MiniTrainingForm, MiniTrainingModelForm, \
-    EducationGroupModelForm, GroupForm
+from base.forms.education_group.create import CreateEducationGroupYearForm, EducationGroupModelForm, GroupForm
+from base.forms.education_group.mini_training import MiniTrainingModelForm, MiniTrainingForm
 from base.forms.education_group.training import TrainingEducationGroupYearForm, TrainingForm
 from base.models.education_group_year import EducationGroupYear
 from base.models.enums import education_group_categories
